refactor: filter category 위치 변경 및 popup 컴포넌트 변경 방식 적용
